# Perf: deprecate mix_arrays use add_weighted instead