Addresses #1606

Refactor `add_weighted` to use cv2 and replace all usages of `mix_arrays` with `add_weighted`. In addition, deprecate the use of `mix_arrays`.

I made a micro-benchmark just to verify that the new function works faster as expected:
**Setup code** 
```python
import numpy as np

from albumentations.augmentations.functional import add_weighted

np.random.seed(0)
SIZE = 1000
img1 = np.random.random((SIZE, SIZE))
img2 = np.random.random((SIZE, SIZE))
```
**Current version**
```python
%%timeit
add_weighted(img1, 0.5, img2, 0.5)
# 7.7 ms ± 362 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

**This PR**
```python
%%timeit
add_weighted(img1, 0.5, img2, 0.5)
# 5.7 ms ± 56.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```